requirements.txt
pandas
numpy
seaborn
scikit-learn
tensorflow
pyyaml
# -e .